Feature/implementation of crud api

+++ src/main/java/com/iop/productservice/repository/ProductRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     // Named query
-    Optional<Product> findByProductIdAndIsDeleted(Integer id, boolean isDeleted);
+    Optional<Product> findByProductIdAndIsDeleted(Long id, boolean isDeleted);
 }
